+++ src/bin/cmds/build.rs
@@ -1,4 +1,4 @@
-use super::{args, match_backends, match_logger, match_threads};
+use super::{args, match_backends, match_idris_opts, match_logger, match_threads};
 use clap::{App, Arg, ArgMatches, SubCommand};
 use elba::{
     cli::build,
@@ -20,6 +20,7 @@ pub fn cli() -> App<'static, 'static> {
         .arg(args::build_threads())
         .arg(args::offline())
         .arg(args::debug_log())
+        .arg(args::idris_opts())
         .args(&args::backends())
 }
 
@@ -37,6 +38,7 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
         threads: match_threads(c, args),
         shell: c.shell(),
         offline: args.is_present("offline"),
+        opts: match_idris_opts(c, args),
     };
 
     let ts = (