+++ src/bin/cmds/doc.rs
@@ -1,4 +1,4 @@
-use super::{args, match_logger, match_threads};
+use super::{args, match_logger, match_idris_opts, match_threads};
 use clap::{App, ArgMatches, SubCommand};
 use elba::{
     cli::build,
@@ -13,6 +13,7 @@ pub fn cli() -> App<'static, 'static> {
         .arg(args::build_threads())
         .arg(args::debug_log())
         .arg(args::offline())
+        .arg(args::idris_opts())
 }
 
 pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
@@ -29,6 +30,7 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
         threads: match_threads(c, args),
         shell: c.shell(),
         offline: args.is_present("offline"),
+        opts: match_idris_opts(c, args),
     };
 
     build::doc(&ctx, &project)