+++ src/bin/cmds/mod.rs
@@ -21,7 +21,7 @@ use itertools::Itertools;
 use slog::{Discard, Drain, Logger};
 use slog_async;
 use slog_term;
-use std::process::Command;
+use std::{env, process::Command};
 
 pub type Exec = fn(&mut Config, &ArgMatches) -> Res<String>;
 
@@ -138,6 +138,20 @@ pub fn match_threads(_c: &mut Config, args: &ArgMatches) -> u32 {
         .unwrap_or(1)
 }
 
+pub fn match_idris_opts(_c: &mut Config, args: &ArgMatches) -> Vec<String> {
+    let mut res = vec![];
+    
+    if let Ok(val) = env::var("IDRIS_OPTS") {
+        res.extend(val.split(' ').map(|x| x.to_string()));
+    }
+    
+    if let Some(vals) = args.values_of("idris-opts") {
+        res.extend(vals.map(|x| x.to_string()));
+    }
+    
+    res
+}
+
 mod args {
     use clap;
 
@@ -216,5 +230,11 @@ mod args {
         Arg::with_name("debug-log")
             .long("debug-log")
             .help("Print debug logs instead of prettified output")
+    }
+    
+    pub fn idris_opts() -> Arg {
+        Arg::with_name("idris-opts")
+            .last(true)
+            .min_values(0)
     }
 }