+++ src/bin/cmds/test.rs
@@ -1,4 +1,4 @@
-use super::{args, match_backends, match_logger, match_threads};
+use super::{args, match_idris_opts, match_backends, match_logger, match_threads};
 use clap::{App, Arg, ArgMatches, SubCommand};
 use elba::{
     cli::build,
@@ -14,6 +14,7 @@ pub fn cli() -> App<'static, 'static> {
         .arg(args::build_threads())
         .arg(args::offline())
         .arg(args::debug_log())
+        .arg(args::idris_opts())
         .arg(
             Arg::with_name("test-threads")
                 .long("test-threads")
@@ -41,6 +42,7 @@ pub fn exec(c: &mut Config, args: &ArgMatches) -> Res<String> {
         threads: match_threads(c, args),
         shell: c.shell(),
         offline: args.is_present("offline"),
+        opts: match_idris_opts(c, args),
     };
 
     // This is where our default codegen backend is set