+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Make elba check for literate Idris files when looking for modules
   (#28).
 
+- Allow passing options to the Idris compiler with `--`.
+
 ## [0.2.0]
 
 This is a major release of elba which polishes up the 0.1.x release